+++ src/db_adapter.h
@@ -23,18 +23,18 @@ public:
     explicit DbAdapter(QDir *decks_path, QString deck_name, QObject *parent = nullptr);
     
     qlonglong newDeckRow();
-    qlonglong newAudioRow(int deck_rowid);
+    qlonglong newAudioRow(qlonglong deck_rowid);
     QList<QMap<QString,QVariant>> selectDeckItems();
     QList<QMap<QString,QVariant>> selectDeckItemsFiltered(QString filter);
     QList<QMap<QString,QVariant>> selectDeckDirtyDozenItems();
-    QList<QMap<QString,QVariant>> selectDeckItem(int rowid);
-    void updateDeckItem(int rowid, QString name, QString word, QString phonetical, QString translation);
-    QList<QMap<QString, QVariant> > deleteItem(int rowid);
-    void deleteImage(int rowid);
+    QList<QMap<QString,QVariant>> selectDeckItem(qlonglong rowid);
+    void updateDeckItem(qlonglong rowid, QString name, QString word, QString phonetical, QString translation);
+    QList<QMap<QString, QVariant> > deleteItem(qlonglong rowid);
+    void deleteImage(qlonglong rowid);
     QList<QMap<QString,QVariant>> audioFilenamesForDeckRowID(qlonglong rowid);
-    void deleteAudio(int rowid);
-    void insertAudioFilename(qlonglong deck_rowid, int audio_rowid, QString filename, QString description);
-    void insertImageFilename(int rowid, QString filename);
+    void deleteAudio(qlonglong rowid);
+    void insertAudioFilename(qlonglong deck_rowid, qlonglong audio_rowid, QString filename, QString description);
+    void insertImageFilename(qlonglong rowid, QString filename);
 };
 
 #endif // DB_ADAPTER_H
